% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh3d.R
\name{as.mesh3d}
\alias{as.mesh3d}
\alias{as.mesh3d.SC0}
\alias{as.mesh3d.SC}
\title{Convert to mesh object}
\usage{
\method{as.mesh3d}{SC0}(x, type = "segments", ...)

\method{as.mesh3d}{SC}(x, ...)
}
\arguments{
\item{x}{}

\item{points_only}{shortcut primitives and just give points}
}
\description{
Converting silicate forms to segment types.
}
\examples{
x <-   silicate::SC0(minimal_mesh)
}
