% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh3d.R
\name{as.mesh3d}
\alias{as.mesh3d}
\alias{as.mesh3d.SC0}
\alias{as.mesh3d.SC}
\alias{as.mesh3d.TRI0}
\alias{as.mesh3d.TRI}
\title{Convert to mesh object}
\usage{
\method{as.mesh3d}{SC0}(x, type = "segments", ...)

\method{as.mesh3d}{SC}(x, ...)

\method{as.mesh3d}{TRI0}(x, type = "segments", ...)

\method{as.mesh3d}{TRI}(x, ...)
}
\arguments{
\item{x}{}

\item{type}{'points' or 'segments'}
}
\description{
Converting silicate forms to segment types.
}
\examples{
x <-   silicate::SC0(minimal_mesh)
x <-   silicate::SC0(minimal_mesh)
}
